+++ src/Homophonic-Alternatives.cpp
@@ -1364,7 +1364,7 @@ Vocinity::Homophonic_Alternative_Composer::get_alternatives(const std::string& r
             tbb::blocked_range<int>(0, words.size()),
             [&](tbb::blocked_range<int> r)
             {
-                for(ushort i = r.begin(); i < r.end(); ++i)
+                for(uint i = r.begin(); i < r.end(); ++i)
                 {
                     const auto& word = words.at(i);
                     if(akil::memory::vector_contains(instructions.dismissed_word_indices, i))
@@ -1386,7 +1386,7 @@ Vocinity::Homophonic_Alternative_Composer::get_alternatives(const std::string& r
     else
 #endif
     {
-        for(ushort order = 0; order < words.size(); ++order)
+        for(uint order = 0; order < words.size(); ++order)
         {
             const auto& word = words.at(order);
             if(akil::memory::vector_contains(instructions.dismissed_word_indices, order))
